requirements.txt
# Core dependencies
gradio>=4.0.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
joblib>=1.3.0

# Weather API
requests>=2.31.0
python-dotenv>=1.0.0

# Optional - for data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
